+++ wp-content/plugins/job-aggregator/config/sources.php
@@ -53,7 +53,7 @@ return [
             ],
         ],
         [
-            "enabled" => false,
+            "enabled" => true,
             "key" => "hotnigerianjobs",
             "driver" => "hotnigerianjobs",
             "label" => "Hot Nigerian Jobs",